Repository reorganization
Reorganized the repo a bit. I've added a top-level cpp folder for C++
code, so that we can more easily accomodate non-C++ headers in the
same repository. I've also added a psu- prefix to the names of the
sub-directories within cpp/include. This is to avoid name conflicts
when using

-Iexternal/psudb-common/cpp/include (or similar)

to add the common headers to a project. Seems like names like ds, util,
and dist, would be pretty obvious candidates for use in the projects
themselves, so now the psudb stuff is clearer,

etc.